fix what the branch review found
The root of the two worst ones: FeatureExecutor::suspend wrapped
FlowStoppedException into a task failure, though execNoResult beside it re-throws
it and .ai/README names suspend as a place that does. A stopped coroutine
therefore reported a broker error, and nothing downstream could tell a deliberate
stop from a dead channel.

On that footing the consumer teardown works as it was meant to. A consumer whose
queue is taken away no longer fails the whole group — which used to cut every
other handler mid-message — and a channel is no longer closed from a coroutine the
scheduler has already detached, where the close suspended forever and left the
channel open on the broker for half an hour.

Data next. A decimal significand at or above 2^31 came back sign-extended and
threw out of the envelope constructor, so the message was redelivered forever; a
timestamp past PHP_INT_MAX wrapped into a date before 1970. A get that outran its
deadline dropped the message the broker had already handed over, and a consume
that outran its left the broker feeding a consumer nobody would ever read.
Confirms and returns queued for a wait loop that never came grew without a bound.

The consumer counters counted ack commands as if they were deliveries, folded
auto-acknowledged ones into the average as zero, and were weighted by a
denominator Go never used. A numeric group name crashed every panel request.

The master replaced a group by the position it held in the file just read rather
than by name, which dropped a different pool from the running list.

+++ ext/internal/features/amqp/values.go
@@ -173,9 +173,13 @@ func fromTableValue(value any) any {
 		}
 	case amqp091.Decimal:
 		return map[string]any{
-			taggedKind:        taggedDecimal,
-			taggedExponent:    int64(typed.Scale),
-			taggedSignificand: int64(typed.Value),
+			taggedKind:     taggedDecimal,
+			taggedExponent: int64(typed.Scale),
+			// Through uint32, not straight from the int32 the driver holds: an AMQP
+			// decimal's significand is unsigned, so anything at or above 2^31 has its
+			// high bit set and a plain int64() would sign-extend it into a negative
+			// number PHP then refuses to build an AMQPDecimal from.
+			taggedSignificand: int64(uint32(typed.Value)),
 		}
 	case []byte:
 		return string(typed)
